pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-checkpoint-lmdb"
version = "0.1.0"
description = "LangGraph checkpointer using LMDB"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "langgraph-checkpoint>=2.0.0",
    "lmdb>=1.4.1",
    "orjson>=3.9.15",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "langchain-openai>=0.1.0",
]
